This is a app that helps optimise data by sending the data to server and allowing the users to access it after completion.
For running the above app one needs to install python 3.0 and above,
also the flask library which can be done by following the link:
http://flask.pocoo.org/docs/0.10/installation/

then open cmd and run the same over python.

This will run on localhost:5000 which you can access on any browser and
the links will be downloaded in the static folder of the app.

You can run the app over local area network if you have many devices
connected to same network and all the data will be downloaded by the
server.

The app is mobile friendly so have fun.